Moved MultiPicker{Media}Type building to a seperate file
